The base class for Sign up and Login forms.
